[MCUX-76833] fix: prevent integer truncation in platform_compute_crc_over_shmem_struct
Changed length variable type from uint16_t to uint32_t in platform_compute_crc_over_shmem_struct
to avoid potential data loss when calculating offset between structure members.
Addresses CERT-C rule INT31-C violation (CID 35174378) by ensuring integer conversions
don't result in lost or misinterpreted data.

+++ lib/rpmsg_lite/porting/platform/mcxw72x/rpmsg_platform.c
@@ -431,8 +431,8 @@ static uint16_t platform_compute_crc_over_shmem_struct(rpmsg_platform_shmem_conf
     uint8_t byte = 0U;
 
     uint8_t *ptr = (uint8_t *)(&protec_shmem_struct->config);
-    uint16_t len = (uint16_t)((uint32_t)(uint8_t *)(&protec_shmem_struct->shmemConfigCrc) -
-                              (uint32_t)(uint8_t *)(&protec_shmem_struct->config));
+    uint32_t len = ((uint32_t)(uint8_t *)(&protec_shmem_struct->shmemConfigCrc) -
+                    (uint32_t)(uint8_t *)(&protec_shmem_struct->config));
     while (len != 0U)
     {
         byte = *ptr;
